Sync with master

+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/types/StructureType.java
@@ -30,5 +30,5 @@ public interface StructureType extends AnnotatableType {
 
     Map<String, Field> getFields();
 
-    int getFlags();
+    long getFlags();
 }
